pyproject.toml
[project]
name = "markdown-demo"
version = "0.1.0"
description = "Operations issue metrics Streamlit app"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = ["python-dotenv (>=1.2.2,<2.0.0)", "streamlit (>=1.55.0,<2.0.0)", "sqlalchemy (>=2.0.48,<3.0.0)", "pg8000 (>=1.31.5,<2.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.13"
mypy_path = "."
files = ["app", "tests"]
pretty = true
show_error_codes = true
warn_unused_ignores = true
no_implicit_optional = true
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest-playwright (>=0.7.2,<0.8.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]
